+++ src/Service/S3/src/Enum/BucketLocationConstraint.php
@@ -15,6 +15,7 @@ final class BucketLocationConstraint
     public const AP_NORTHEAST_3 = 'ap-northeast-3';
     public const AP_SOUTHEAST_1 = 'ap-southeast-1';
     public const AP_SOUTHEAST_2 = 'ap-southeast-2';
+    public const AP_SOUTHEAST_3 = 'ap-southeast-3';
     public const AP_SOUTH_1 = 'ap-south-1';
     public const CA_CENTRAL_1 = 'ca-central-1';
     public const CN_NORTHWEST_1 = 'cn-northwest-1';
@@ -44,6 +45,7 @@ final class BucketLocationConstraint
             self::AP_NORTHEAST_3 => true,
             self::AP_SOUTHEAST_1 => true,
             self::AP_SOUTHEAST_2 => true,
+            self::AP_SOUTHEAST_3 => true,
             self::AP_SOUTH_1 => true,
             self::CA_CENTRAL_1 => true,
             self::CN_NORTHWEST_1 => true,