+++ src/Service/S3/src/S3Client.php
@@ -1014,6 +1014,7 @@ class S3Client extends AbstractApi
             case 'ap-southeast-3':
             case 'ca-central-1':
             case 'eu-central-1':
+            case 'eu-central-2':
             case 'eu-north-1':
             case 'eu-south-1':
             case 'eu-west-1':