+++ src/Service/Kms/src/KmsClient.php
@@ -321,6 +321,7 @@ class KmsClient extends AbstractApi
             case 'ap-northeast-2':
             case 'ap-northeast-3':
             case 'ap-south-1':
+            case 'ap-south-2':
             case 'ap-southeast-1':
             case 'ap-southeast-2':
             case 'ap-southeast-3':