+++ src/Service/S3/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Added `me-central-1`, `ap-southeast-3`, `eu-central-2` and `eu-south-2` regions
+- Added `me-central-1`, `ap-southeast-3`, `eu-central-2`, `eu-south-2` and `ap-south-2` regions
 
 ## 1.12.0